refactor: refactoring AddFoodModal
