--- src/Components/AddFoodModal/types.ts
@@ -1,4 +1,0 @@
-export type AddFoodModalProps = {
-  modalVisible: boolean;
-  setModalVisible: (value: boolean) => void;
-};
