+++ src/Components/AddStory/index.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { Text } from "react-native";
+import { View } from "react-native-reanimated/lib/typescript/Animated";
+
+const AddStory = () => {
+  return (
+    <View>
+      <Text>adding somethings</Text>
+    </View>
+  );
+};
+
+export default AddStory;
