+++ src/Components/WithModal/index.tsx
@@ -1,13 +1,17 @@
 // libs
 import React from "react";
-import { Text, View } from "react-native";
+import { View } from "react-native";
 import { Modal, ModalContent, SlideAnimation } from "react-native-modals";
 
 // custom
-import { AddFoodModalProps } from "./types";
+import { WithModalProps } from "./types";
 import { styles } from "./styles";
 
-const AddFoodModal = ({ modalVisible, setModalVisible }: AddFoodModalProps) => {
+const WithModal = ({
+  modalVisible,
+  setModalVisible,
+  children,
+}: WithModalProps) => {
   return (
     <Modal
       visible={modalVisible}
@@ -21,14 +25,10 @@ const AddFoodModal = ({ modalVisible, setModalVisible }: AddFoodModalProps) => {
       modalAnimation={new SlideAnimation({ slideFrom: "bottom" })}
     >
       <ModalContent>
-        <View style={styles.modalCtr}>
-          <View style={styles.horizontalLine} />
-          <View></View>
-          <Text>something</Text>
-        </View>
+        <View style={styles.modalCtr}>{children}</View>
       </ModalContent>
     </Modal>
   );
 };
 
-export default AddFoodModal;
+export default WithModal;
