+++ src/Components/WithModal/types.ts
@@ -1,0 +1,7 @@
+import React from "react";
+
+export type WithModalProps = {
+  modalVisible: boolean;
+  setModalVisible: (value: boolean) => void;
+  children: React.ReactElement;
+};
